Fix a bug in get_sample_file_peak_timeseries
After some dependency update, it had started raising exception related to numpy.float32 datatype. Convert to native float to fix it.

+++ backend/backend/api_rest/controllers/sample_files_controller.py
@@ -236,6 +236,6 @@ async def get_sample_file_peak_timeseries(
 
     return {
         "mz": peak_mz_data,
-        "intensity": list(peak_timeseries.values),
-        "time": list(peak_timeseries.time.values),
+        "intensity": list(peak_timeseries.values.astype(float)),
+        "time": list(peak_timeseries.time.values.astype(float)),
     }
